refactor: consolidate clases required for extensions
