+++ src/components/Observers/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { Card, Button, TextInput, FAB, Portal, useTheme } from 'react-native-paper';
 import { Picker } from '@react-native-picker/picker';
-import { observersComponentBuilderMap } from '@/src/components/Observers/helpers/observersComponentBuilderMap';
-import { observersTitleBuilderMap } from '@/src/components/Observers/helpers/observersTitleBuilderMap';
+import { observersComponentBuilderMap } from '@/src/extensions/observersComponentBuilderMap';
+import { observersTitleBuilderMap } from '@/src/extensions/observersTitleBuilderMap';
 import { ObserversParams } from '@/src/types/ObserversParams';
 
 export const Observers = ({ observables, add, remove, update, save }: ObserversParams) => {
